Move getOrCompute to a new Cache module
