+++ lib/template-dependencies.js
@@ -2,11 +2,11 @@
  * Credit goes to @zwilias, from his PR here https://github.com/rtfeldman/node-test-runner/pull/356/files
  */
 
-const path = require('path');
 const util = require('util');
 const fs = require('fs-extra');
 const spawnAsync = require('cross-spawn');
 const Hash = require('./hash');
+const Cache = require('./cache');
 const ErrorMessage = require('./error-message');
 const getExecutable = require('elm-tooling/getExecutable');
 
@@ -17,33 +17,19 @@ module.exports = {
   addElmSyntax
 };
 
-const fsEnsureDir = util.promisify(fs.ensureDir);
 const fsReadJson = util.promisify(fs.readJson);
 const fsWriteJson = util.promisify(fs.writeJson);
 
 // GET
 
-async function getFromCacheOrCompute(folder, key, fn) {
-  const filepath = path.join(folder, `${key}.json`);
-
-  let cachedResult = await fsReadJson(filepath).catch(() => null);
-  if (cachedResult) {
-    return cachedResult;
-  }
-
-  const result = await fn();
-
-  fsEnsureDir(folder).then(() =>
-    fsWriteJson(filepath, result)
-  );
-
-  return result;
-}
-
 async function get(options, elmJsonDependencies, pathToElmJson) {
   const dependencyHash = Hash.hash(JSON.stringify(elmJsonDependencies));
-  const cacheKey = `${dependencyHash}${options.localElmReviewSrc ? '-local' : ''}`;
-  return getFromCacheOrCompute(options.dependenciesCachePath(), cacheKey, () => computeDependencies(options, pathToElmJson));
+  const cacheKey = `${dependencyHash}${
+    options.localElmReviewSrc ? '-local' : ''
+  }`;
+  return Cache.getOrCompute(options.dependenciesCachePath(), cacheKey, () =>
+    computeDependencies(options, pathToElmJson)
+  );
 }
 
 function computeDependencies(options, pathToElmJson) {
@@ -75,6 +61,7 @@ ${formatElmJsonError(error, options)}`,
         delete dependencies.direct['jfmengels/elm-review'];
         delete dependencies.indirect['jfmengels/elm-review'];
       }
+
       return dependencies;
     });
 }
